+++ admin/src/components/PluginIcon.jsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import { Search } from '@strapi/icons';
 
 const PluginIcon = () => <Search />;
 
-export default PluginIcon;
+export { PluginIcon };